More task queue tweaks; optimize TaskContainer.TasksRunning next
